Attach response to request in TransitLogger (#2)
The response status in pyramid sawing log is always "200 OK" even when
the response is not 200.  The reason for this seems to be in
`TransitLogger` we generated a response locally, but in `EnvironFilter`
we get the response using `request.response` which is not the same as
the one in `TransitLogger`.

Add `request.response = response` in `TransitLogger` to make sure the
response is correct in `EnvironFilter`.